requirements.txt
fastapi
uvicorn
openai
langchain
python-dotenv
typer 
rich
streamlit
langchain 
faiss-cpu 
openai
